RP-283: updates based on code review
