+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/ReadOnlyExamRepository.java
@@ -1,0 +1,17 @@
+package org.opentestsystem.rdw.ingest.processor.repository;
+
+import org.opentestsystem.rdw.ingest.processor.model.Exam;
+
+/**
+ * Repository for managing {@link Exam}
+ */
+public interface ReadOnlyExamRepository {
+    /**
+     * Finds an exam by its unique identifiers
+     *
+     * @param assessmentId  an assessment if
+     * @param opportunityId an opportunity id
+     * @return the found exam, null if not found
+     */
+    Exam findOne(long assessmentId, String opportunityId);
+}
