+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/service/AssessmentService.java
@@ -1,6 +1,5 @@
 package org.opentestsystem.rdw.ingest.processor.service;
 
-import org.opentestsystem.rdw.common.model.trt.Test;
 import org.opentestsystem.rdw.ingest.processor.model.Assessment;
 
 /**
@@ -9,10 +8,10 @@ import org.opentestsystem.rdw.ingest.processor.model.Assessment;
 public interface AssessmentService {
 
     /**
-     * Finds an assessment definition for the given {@link Test}
+     * Finds an assessment definition for the given natural ID
      *
-     * @param test the {@link Test}
-     * @return the {@link Assessment}, null if not found
+     *
+     * @param naturalId@return the {@link Assessment}, null if not found
      */
-    Assessment findOneForTest(Test test);
+    Assessment findOneForNaturalId(final String naturalId);
 }
