+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/service/impl/DefaultAssessmentService.java
@@ -3,7 +3,6 @@ package org.opentestsystem.rdw.ingest.processor.service.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import org.opentestsystem.rdw.common.model.trt.Test;
 import org.opentestsystem.rdw.ingest.processor.model.Assessment;
 import org.opentestsystem.rdw.ingest.processor.repository.AssessmentRepository;
 import org.opentestsystem.rdw.ingest.processor.service.AssessmentService;
@@ -18,7 +17,7 @@ class DefaultAssessmentService implements AssessmentService {
     }
 
     @Override
-    public Assessment findOneForTest(final Test test) {
-        return assessmentRepository.findOneByNaturalId(test.getName());
+    public Assessment findOneForNaturalId(final String naturalId) {
+        return assessmentRepository.findOneByNaturalId(naturalId);
     }
 }
