+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/service/impl/DefaultTDSReportProcessor.java
@@ -48,9 +48,10 @@ class DefaultTDSReportProcessor implements TDSReportProcessor {
             final Integer schoolId = examineeProcessor.parseSchoolId(examinee);
             if (schoolId == null) throw new IllegalArgumentException("School Id may not be null");
 
-            final Assessment assessment = assessmentService.findOneForTest(report.getTest());
-            if (assessment == null) throw new IllegalArgumentException("Assessment may not be null");
+            if (report.getTest() == null) throw new IllegalArgumentException("Test cannot be null");
 
+            final Assessment assessment = assessmentService.findOneForNaturalId(report.getTest().getName());
+            if (assessment == null) throw new IllegalArgumentException("Assessment may not be null");
 
             final Exam exam = studentExamProcessor.parseExam(report, schoolId, assessment);
             final Student student = examineeProcessor.parseStudent(examinee, exam.getSchoolYear(), schoolId);
