+++ exam-processor/src/main/resources/scripts/ExamPipeline.json
@@ -3,7 +3,6 @@
   "version": "1.0",
   "type": "PIPELINE",
   "rollbackVersion": "",
-  "options": ["xml"],
   "nodeDefinitions": [
     {
       "name": "trt_transform",
