+++ exam-processor/src/test/java/org/opentestsystem/rdw/ingest/processor/ExamProcessorTransformationTest.java
@@ -94,7 +94,6 @@ public class ExamProcessorTransformationTest {
                 .version("1")
                 .rollbackVersion("")
                 .nodeDefinitions(singletonList(nodeDefinition))
-                .options(singletonList("xml"))
                 .build();
 
         ScriptSource scriptSource = mock(ScriptSource.class);
