+++ exam-processor/src/test/java/org/opentestsystem/rdw/ingest/processor/ExamProcessorValidationTest.java
@@ -1,5 +1,6 @@
 package org.opentestsystem.rdw.ingest.processor;
 
+import org.apache.commons.io.IOUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.core.io.ResourceLoader;
@@ -7,8 +8,6 @@ import org.springframework.http.MediaType;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.GenericMessage;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
@@ -95,7 +94,6 @@ public class ExamProcessorValidationTest {
                 .version("1")
                 .rollbackVersion("")
                 .nodeDefinitions(singletonList(nodeDefinition))
-                .options(singletonList("xml"))
                 .build();
 
         ScriptSource scriptSource = mock(ScriptSource.class);
@@ -250,7 +248,7 @@ public class ExamProcessorValidationTest {
 
     private byte [] loadFromResourceFile(final String name) throws Exception {
         InputStream is = this.getClass().getResourceAsStream(name);
-        return readFully(is);
+        return IOUtils.toByteArray(is);
     }
 
     private String loadStringFromResourceFile(String name) throws Exception {
@@ -261,21 +259,4 @@ public class ExamProcessorValidationTest {
         final String scriptCode = loadStringFromResourceFile(filePath);
         return new ScriptInfo(scriptCode, properties);
     }
-
-    private static byte[] readFully(InputStream in) throws IOException {
-
-        ByteArrayOutputStream os = new ByteArrayOutputStream();
-
-        byte[] buffer = new byte[1024];
-        int len;
-
-        // read bytes from the input stream and store them in buffer
-        while ((len = in.read(buffer)) != -1) {
-            // write bytes from the buffer into output stream
-            os.write(buffer, 0, len);
-        }
-
-        return os.toByteArray();
-    }
-
 }
