+++ exam-processor/src/test/java/org/opentestsystem/rdw/ingest/processor/service/impl/DefaultAssessmentServiceTest.java
@@ -43,7 +43,7 @@ public class DefaultAssessmentServiceTest {
         when(test.getAssessmentType()).thenReturn("testAsmtType");
 
         try {
-            service.findOneForTest(test);
+            service.findOneForNaturalId(test.getName());
         } catch (final ImportException ex) {
             assertThat(ex.getStatus()).isEqualTo(BAD_DATA);
             assertThat(ex.getMessage()).isEqualTo("{\"messages\":[" +
@@ -67,6 +67,6 @@ public class DefaultAssessmentServiceTest {
         when(test.getSubject()).thenReturn("testSubject");
         when(test.getAssessmentType()).thenReturn("Interim");
 
-        assertThat(service.findOneForTest(test).getNaturalId()).isEqualTo(asmtNaturalId);
+        assertThat(service.findOneForNaturalId(test.getName()).getNaturalId()).isEqualTo(asmtNaturalId);
     }
 }
