+++ exam-processor/src/test/java/org/opentestsystem/rdw/ingest/processor/service/impl/DefaultTDSReportProcessorTest.java
@@ -2,6 +2,8 @@ package org.opentestsystem.rdw.ingest.processor.service.impl;
 
 import org.junit.Before;
 import org.junit.Test;
+
+import java.text.ParseException;
 import org.opentestsystem.rdw.common.model.trt.Examinee;
 import org.opentestsystem.rdw.common.model.trt.Opportunity;
 import org.opentestsystem.rdw.common.model.trt.TDSReport;
@@ -14,8 +16,6 @@ import org.opentestsystem.rdw.ingest.processor.service.ExamineeProcessor;
 import org.opentestsystem.rdw.ingest.processor.service.StudentExamProcessor;
 import org.opentestsystem.rdw.ingest.processor.service.StudentExamWriter;
 import org.opentestsystem.rdw.ingest.processor.service.TDSReportProcessor;
-
-import java.text.ParseException;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -66,7 +66,7 @@ public class DefaultTDSReportProcessorTest {
 
         final Assessment assessment = mock(Assessment.class);
         when(assessment.getSubjectId()).thenReturn(1);
-        when(assessmentService.findOneForTest(test)).thenReturn(assessment);
+        when(assessmentService.findOneForNaturalId(test.getName())).thenReturn(assessment);
         when(studentExamProcessor.parseExam(tdsReport, 99, assessment)).thenReturn(exam);
 
         final Student student = mock(Student.class);
