+++ script-common/src/main/java/org/opentestsystem/rdw/ingest/script/PipelineDefinition.java
@@ -1,6 +1,5 @@
 package org.opentestsystem.rdw.ingest.script;
 
-import java.util.Collections;
 import java.util.List;
 
 /**
@@ -8,7 +7,6 @@ import java.util.List;
  */
 public class PipelineDefinition extends VersionedResource {
     private final String rollbackVersion;
-    private final List<String> options;
     private final List<VersionedResource> nodeDefinitions;
 
     /**
@@ -17,11 +15,9 @@ public class PipelineDefinition extends VersionedResource {
     private PipelineDefinition(final String name,
                                final String version,
                                final String rollbackVersion,
-                               final List<String> options,
                                final List<VersionedResource> nodeDefinitions) {
         super(name, version, ResourceType.PIPELINE);
         this.rollbackVersion = rollbackVersion;
-        this.options = options;
         this.nodeDefinitions = nodeDefinitions;
     }
 
@@ -33,10 +29,6 @@ public class PipelineDefinition extends VersionedResource {
         return rollbackVersion;
     }
 
-    public List<String> getOptions() {
-        return options;
-    }
-
     public List<VersionedResource> getNodeDefinitions() {
         return nodeDefinitions;
     }
@@ -45,7 +37,6 @@ public class PipelineDefinition extends VersionedResource {
         private String name;
         private String version;
         private String rollbackVersion;
-        private List<String> options = Collections.emptyList();
         private List<VersionedResource> nodeDefinitions;
 
         PipelineDefinitionBuilder() {
@@ -66,11 +57,6 @@ public class PipelineDefinition extends VersionedResource {
             return this;
         }
 
-        public PipelineDefinitionBuilder options(final List<String> options) {
-            this.options = options;
-            return this;
-        }
-
         public PipelineDefinitionBuilder nodeDefinitions(final List<VersionedResource> nodeDefinitions) {
             this.nodeDefinitions = nodeDefinitions;
             return this;
@@ -88,24 +74,19 @@ public class PipelineDefinition extends VersionedResource {
             this.rollbackVersion = rollbackVersion;
         }
 
-        public void setOptions(final List<String> options) {
-            this.options = options;
-        }
-
         public void setNodeDefinitions(final List<VersionedResource> nodeDefinitions) {
             this.nodeDefinitions = nodeDefinitions;
         }
 
         public PipelineDefinition build() {
-            return new PipelineDefinition(name, version, rollbackVersion, options, nodeDefinitions);
+            return new PipelineDefinition(name, version, rollbackVersion, nodeDefinitions);
         }
 
         public String toString() {
             return "PipelineDefinition.PipelineDefinitionBuilder(name="
                     + this.name + ", version="
                     + this.version + ", rollbackVersion="
-                    + this.rollbackVersion + ", options="
-                    + this.options + ", nodeDefinitions="
+                    + this.rollbackVersion  + ", nodeDefinitions="
                     + this.nodeDefinitions + ")";
         }
     }
