+++ script-common/src/main/java/org/opentestsystem/rdw/ingest/script/PipelineScript.java
@@ -26,6 +26,4 @@ public abstract class PipelineScript extends Script {
         setProperty("input", input);
         return this;
     }
-
-    abstract public PipelineScript enableXmlExtensions();
 }
