+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/DefaultScriptSourceTest.java
@@ -34,6 +34,6 @@ public class DefaultScriptSourceTest {
     @Test
     public void isShouldLoadScript() {
         final VersionedResource node = new VersionedResource("exam_script", "", ResourceType.USER_NODE);
-        assertThat(scriptSource.loadScript(node).getCode()).startsWith("transform '//Item'");
+        assertThat(scriptSource.loadScript(node).getCode()).startsWith("enable 'xml'");
     }
 }
