+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/PipelineDefinitionJsonTest.java
@@ -24,7 +24,6 @@ public class PipelineDefinitionJsonTest {
         pipelineDefinition = PipelineDefinition.builder()
                 .name("Exam Pipeline")
                 .version("1.0")
-                .options(singletonList("xml"))
                 .nodeDefinitions(singletonList(nodeDefinition))
                 .build();
     }
@@ -40,7 +39,6 @@ public class PipelineDefinitionJsonTest {
 
         assertThat(newPipelineDefinition.getName(), is(pipelineDefinition.getName()));
         assertThat(newPipelineDefinition.getVersion(), is(pipelineDefinition.getVersion()));
-        assertThat(newPipelineDefinition.getOptions(), is(pipelineDefinition.getOptions()));
         assertThat(newPipelineDefinition.getNodeDefinitions().size(), is(1));
         assertThat(newPipelineDefinition.getNodeDefinitions(), hasItem(nodeDefinition));
     }
