+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/PipelineDefinitionTest.java
@@ -15,7 +15,6 @@ public class PipelineDefinitionTest {
                 .name("test")
                 .version("23")
                 .rollbackVersion("16")
-                .options(ImmutableList.of("xml"))
                 .nodeDefinitions(ImmutableList.of(node))
                 .build();
 
@@ -23,7 +22,6 @@ public class PipelineDefinitionTest {
         assertThat(pipelineDefinition.getName()).isEqualTo("test");
         assertThat(pipelineDefinition.getVersion()).isEqualTo("23");
         assertThat(pipelineDefinition.getRollbackVersion()).isEqualTo("16");
-        assertThat(pipelineDefinition.getOptions()).containsExactly("xml");
         assertThat(pipelineDefinition.getNodeDefinitions()).containsExactly(node);
     }
 
@@ -32,10 +30,9 @@ public class PipelineDefinitionTest {
         final PipelineDefinition.PipelineDefinitionBuilder builder = PipelineDefinition.builder()
                 .name("test")
                 .version("23")
-                .rollbackVersion("16")
-                .options(ImmutableList.of("xml"));
+                .rollbackVersion("16");
 
-        assertThat(builder.toString()).contains("test").contains("23").contains("16").contains("xml");
+        assertThat(builder.toString()).contains("test").contains("23").contains("16");
     }
 
     @Test
@@ -45,14 +42,12 @@ public class PipelineDefinitionTest {
                 .name("test")
                 .version("23")
                 .rollbackVersion("16")
-                .options(ImmutableList.of("xml"))
                 .nodeDefinitions(ImmutableList.of(node))
                 .build();
         final PipelineDefinition two = PipelineDefinition.builder()
                 .name("test")
                 .version("23")
                 .rollbackVersion("19")
-                .options(ImmutableList.of("csv"))
                 .nodeDefinitions(ImmutableList.of(node))
                 .build();
