+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/PipelineFactoryTest.java
@@ -10,11 +10,9 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 import java.util.Arrays;
 import java.util.Map;
-
 import org.opentestsystem.rdw.ingest.script.impl.DefaultPipelineFactory;
 
 import static java.util.Collections.emptyMap;
-import static java.util.Collections.singletonList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.anyMap;
@@ -52,7 +50,6 @@ public class PipelineFactoryTest {
                 .name("test")
                 .version("1.0")
                 .nodeDefinitions(Arrays.asList(nodeDefinition, nodeDefinition, nodeDefinition))
-                .options(singletonList("xml"))
                 .build();
 
         final String scriptCode = "return 'results'";
