+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/PropertyResolverTest.java
@@ -21,8 +21,8 @@ import static org.opentestsystem.rdw.ingest.script.util.TestUtils.loadStringFrom
 public class PropertyResolverTest {
     private static final String PIPELINE_NAME = "My Test Pipeline";
     private static final String PIPELINE_VERSION = "1.0";
-    private static final String MISSING_PROPERTY_SCRIPT_CODE = "return cat.equals('cat')";
     private static final String SAFE_SCRIPT_CODE = "return 'transformed input'";
+    private static final String MISSING_PROPERTY_SCRIPT_CODE = "cat.equals('cat'); " + SAFE_SCRIPT_CODE;
 
     private PipelineFactory pipelineFactory;
     private ScriptPipelineConfiguration config = new ScriptPipelineConfiguration();
