+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/ScriptCompilerTest.java
@@ -35,7 +35,6 @@ public class ScriptCompilerTest {
     @Before
     public void setup() throws Exception {
         when(scriptSource.loadScriptBaseClass()).thenReturn(loadFromResourceFile("/scripts/DSLScriptBase.groovy"));
-        when(scriptSource.loadScriptBaseClass()).thenReturn(loadFromResourceFile("/scripts/DSLScriptBase.groovy"));
 
         compiler = new ScriptCompiler(configuration.groovyClassLoader(scriptSource));
     }
@@ -67,6 +66,23 @@ public class ScriptCompilerTest {
         Class<? extends Script> scriptClass = compiler.compile("return 1 / 0");
         scriptClass.getConstructor().newInstance().run();
     }
+
+    @Test
+    public void testValidDslEnable() throws Exception {
+        final Class<? extends PipelineScript> scriptClass = compiler.compile("enable 'xml'; outputXml");
+        final String sampleXml = "<sample><field>1</field></sample>";
+        scriptClass.newInstance().bindInput(sampleXml).run();
+    }
+
+    @Test
+    public void testInvalidDslEnable() throws Exception {
+        exception.expect(RuntimeException.class);
+        exception.expectMessage(containsString("bogus type"));
+
+        final Class<? extends PipelineScript> scriptClass = compiler.compile("enable 'bogus type'; return 'hello'");
+        scriptClass.newInstance().run();
+    }
+
 
     private String loadFromResourceFile(String name) throws Exception {
         Reader scriptReader = new InputStreamReader(this.getClass().getResourceAsStream(name));
