+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/ScriptDefinitionTest.java
@@ -76,11 +76,6 @@ public class ScriptDefinitionTest {
         }
 
         @Override
-        public PipelineScript enableXmlExtensions() {
-            return this;
-        }
-
-        @Override
         public Object run() {
             return "Test Results";
         }
