+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/ValidatorScriptTest.java
@@ -1,6 +1,5 @@
 package org.opentestsystem.rdw.ingest.script;
 
-import groovy.lang.Script;
 import org.apache.tools.ant.util.FileUtils;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.Matcher;
@@ -25,6 +24,8 @@ import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ValidatorScriptTest {
+    private static final String SAMPLE_INPUT = "Hello, world!";
+
     @Mock
     private ScriptSource scriptSource;
 
@@ -43,9 +44,9 @@ public class ValidatorScriptTest {
     }
 
     @Test
-    public void itShouldReturnTrueIfNoErrors() {
+    public void itShouldReturnOriginalInputIfNoErrors() {
         final String script = "checkValid";
-        assertThat(runScript(script)).isTrue();
+        assertThat(runScript(script, SAMPLE_INPUT)).isEqualTo(SAMPLE_INPUT);
     }
 
     @Test
@@ -58,7 +59,7 @@ public class ValidatorScriptTest {
         exception.expect(statusMatches(ImportStatus.BAD_DATA));
 
         final String script = "addError('element','value','error');checkValid";
-        runScript(script);
+        runScript(script, "any input");
     }
 
     @Test
@@ -71,7 +72,7 @@ public class ValidatorScriptTest {
         exception.expect(statusMatches(ImportStatus.UNKNOWN_SCHOOL));
 
         final String script = "addError('element','value','error'); checkValid 'UNKNOWN_SCHOOL'";
-        runScript(script);
+        runScript(script, "any input");
     }
 
     private Matcher<Object> statusMatches(ImportStatus status) {
@@ -83,10 +84,10 @@ public class ValidatorScriptTest {
         return FileUtils.readFully(scriptReader);
     }
 
-    private Boolean runScript(final String script) {
+    private Object runScript(final String script, final Object input) {
         try {
-            Class<? extends Script> scriptClass = compiler.compile(script);
-            return (Boolean) scriptClass.getConstructor().newInstance().run();
+            Class<? extends PipelineScript> scriptClass = compiler.compile(script);
+            return scriptClass.getConstructor().newInstance().bindInput(input).run();
         } catch (RuntimeException e) {
             throw e;
         } catch (Exception e) {
