+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/XmlPipelineTest.java
@@ -64,7 +64,6 @@ public class XmlPipelineTest {
                 .version(PIPELINE_VERSION)
                 .rollbackVersion("N/A")
                 .nodeDefinitions(singletonList(nodeDefinition))
-                .options(singletonList("xml"))
                 .build();
 
         when(scriptSource.loadPipelineDefinition(PIPELINE_NAME, PIPELINE_VERSION))
@@ -131,6 +130,7 @@ public class XmlPipelineTest {
         // Test ability to filter out elements with a boolean rule. All elements matching the XPath expression
         // and also matching this rule will be removed.
         final String scriptCode =
+                "enable 'xml'" + LF +
                 "delete '//Item' when { item ->" + LF +
                 "    item.key == '37115' || item.key == '92135'" + LF +
                         "}" + LF +
@@ -155,6 +155,7 @@ public class XmlPipelineTest {
         // Test filter in with a boolean rule. All elements matching the XPath expression, but not matching the rule
         // will be removed.
         String scriptCode =
+                "enable 'xml'" + LF +
                 "delete '//Item' unless { item ->" + LF +
                         "    item.key == '37115'" + LF +
                         "}" + LF +
@@ -193,7 +194,6 @@ public class XmlPipelineTest {
 
         DocumentBuilder db = DocumentBuilderFactory.newInstance().newDocumentBuilder();
         InputSource is = new InputSource(new ByteArrayInputStream(xml));
-//        is.setCharacterStream(new StringReader(xml));
         Document doc = db.parse(is);
