+++ script-common/src/test/java/org/opentestsystem/rdw/ingest/script/util/TestUtils.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.ingest.script.util;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
+import org.apache.commons.io.IOUtils;
+
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
@@ -13,7 +13,7 @@ import org.opentestsystem.rdw.ingest.script.ScriptInfo;
 public class TestUtils {
     public static byte [] loadFromResourceFile(String name) throws Exception {
         InputStream is = TestUtils.class.getResourceAsStream(name);
-        return readFully(is);
+        return IOUtils.toByteArray(is);
     }
 
     public static String loadStringFromResourceFile(String name) throws Exception {
@@ -24,21 +24,4 @@ public class TestUtils {
         final String scriptCode = loadStringFromResourceFile(filePath);
         return new ScriptInfo(scriptCode, properties);
     }
-
-    private static byte[] readFully(InputStream in) throws IOException {
-
-        ByteArrayOutputStream os = new ByteArrayOutputStream();
-
-        byte[] buffer = new byte[1024];
-        int len;
-
-        // read bytes from the input stream and store them in buffer
-        while ((len = in.read(buffer)) != -1) {
-            // write bytes from the buffer into output stream
-            os.write(buffer, 0, len);
-        }
-
-        return os.toByteArray();
-    }
-
 }
