+++ script-common/src/test/resources/scripts/exam_script.groovy
@@ -1,3 +1,5 @@
+enable 'xml'
+
 transform '//Item' by { item ->
     if (item.bankKey.startsWith('10')) {
         item.bankKey = item.bankKey.substring(2)
