+++ script-common/src/test/resources/scripts/exam_script_hybrid.groovy
@@ -1,3 +1,5 @@
+enable 'xml'
+
 applyXsl xslItemRulesOnly
 
 transform '//Response' by { response ->
