+++ script-common/src/test/resources/scripts/exam_script_xsl.groovy
@@ -1,3 +1,5 @@
+enable 'xml'
+
 applyXsl xslAllRules
 
 outputXml
