+++ script-common/src/test/resources/scripts/test_pipeline.json
@@ -3,7 +3,6 @@
   "version": "1.0",
   "type": "PIPELINE",
   "rollbackVersion": "",
-  "options": ["xml"],
   "nodeDefinitions": [{
     "name": "trt_transform",
     "version": "1.0",
